Add deleted column to chats table

+++ db/migrations/20180331_231206_create_express_entry_draw_chats.js
@@ -1,8 +1,7 @@
-
 exports.up = function (knex, Promise) {
   return knex.schema.createTableIfNotExists('express_entry_draw_chats', (table) => {
     table.increments('id').primary()
-    table.integer('chat_id').notNullable()
+    table.bigInteger('chat_id').notNullable()
     table.string('comment').notNullable()
 
     table.index('chat_id')
